Integrate payment with Paypal

+++ db/migrate/20170703165412_create_courses.rb
@@ -1,9 +1,9 @@
- class CreateCourses < ActiveRecord::Migration[5.1]
+class CreateCourses < ActiveRecord::Migration[5.1]
   def change
     create_table :courses do |t|
       t.string :title
       t.text :description
-      t.money :price
+      t.float :price
       t.string :status
       t.string :image
